// src/api/generated/models/bulk-update-request-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Ankieter API
 * Ankieter 6.0 - the API contract, source of truth for the backend and the frontend
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BulkOperationDto } from './bulk-operation-dto';
// May contain unused imports in some cases
// @ts-ignore
import type { QuestionFilterDto } from './question-filter-dto';

/**
 * 
 * @export
 * @interface BulkUpdateRequestDto
 */
export interface BulkUpdateRequestDto {
    /**
     * The questions ticked on screen. Either this or filter.
     * @type {Array<string>}
     * @memberof BulkUpdateRequestDto
     */
    'questionIds'?: Array<string>;
    /**
     * 
     * @type {QuestionFilterDto}
     * @memberof BulkUpdateRequestDto
     */
    'filter'?: QuestionFilterDto;
    /**
     * 
     * @type {BulkOperationDto}
     * @memberof BulkUpdateRequestDto
     */
    'operation': BulkOperationDto;
    /**
     * 
     * @type {Array<string>}
     * @memberof BulkUpdateRequestDto
     */
    'tagIds'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof BulkUpdateRequestDto
     */
    'categoryId'?: string;
    /**
     * 
     * @type {string}
     * @memberof BulkUpdateRequestDto
     */
    'reviewReason'?: string;
    /**
     * Count and sample what would change, changing nothing.
     * @type {boolean}
     * @memberof BulkUpdateRequestDto
     */
    'dryRun'?: boolean;
    /**
     * Required when more than 50 questions change.
     * @type {boolean}
     * @memberof BulkUpdateRequestDto
     */
    'confirmLargeChange'?: boolean;
}